+++ .gitignore
@@ -1,8 +1,8 @@
 .vscode/
 .tox/
-tests/__pycache__/
+__pycache__/
 elixir_rems_proxy/__pycache__/
 elixir_rems_proxy/utils/__pycache__/
 .coverage
 .python-version
-elixir_rems_proxy/utils/*.json
+elixir_rems_proxy/config/*.json
