+++ elixir_rems_proxy/config/__init__.py
@@ -1,14 +1,18 @@
-"""Application Configuration."""
+"""Application configuration."""
 
 import os
 import sys
 import json
+import logging
 
 from pathlib import Path
 from configparser import ConfigParser
 from collections import namedtuple
+from distutils.util import strtobool
 
-from ..utils.logging import LOG
+formatting = '[%(asctime)s][%(name)s][%(process)d %(processName)s][%(levelname)-8s] (L:%(lineno)s) %(module)s | %(funcName)s: %(message)s'
+logging.basicConfig(level=logging.DEBUG if bool(strtobool(os.environ.get('DEBUG', 'False'))) else logging.INFO, format=formatting)
+LOG = logging.getLogger("ELIXIR")
 
 
 def load_json_file(path):
