--- elixir_rems_proxy/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Utilities Module."""
